fix: pass employeeId to projects getter and find corresponding employee
